split previous api.py into api config, question model and routes class. routes and question model completed.
