+++ backend/question.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class Question(BaseModel):
+        question: str
+        session_id: str = None
